add check if signature uploaded before user can sent request - fix upload first time signature  no need to delete old file not exisits

+++ app/Livewire/Profile/Signature.php
@@ -48,9 +48,12 @@ class Signature extends Component
 
 
             if ($this->suser->save()) {
-                if (Storage::disk('signature')->exists($this->current_signature_name)) {
+                if ($this->current_signature_name) {
 
-                    Storage::disk('signature')->delete($this->current_signature_name);
+                    if (Storage::disk('signature')->exists($this->current_signature_name)) {
+
+                        Storage::disk('signature')->delete($this->current_signature_name);
+                    }
                 }
                 $this->current_signature_name = $this->suser->signature;
                 $this->current_signature  = $this->storage2base64(Storage::disk("signature")->get($this->current_signature_name), $this->current_signature_name);
